# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: StewardMesh API
  version: 0.1.0
  description: Requirements-driven inventory and lifecycle management API.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/assets:
    get:
      operationId: listAssets
      tags: [Atlas]
      responses:
        '200':
          description: Asset collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetCollection'
    post:
      operationId: createAsset
      tags: [Atlas]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Asset'
      responses:
        '201':
          description: Asset created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '400':
          description: Invalid request
...
components:
  schemas:
    Asset:
      type: object
      required: [id, name, kind]
      properties:
        id: {type: string}
        name: {type: string}
        kind: {type: string}
        serialNumber: {type: string}
        departmentId: {type: string}
        userId: {type: string}
        status: {type: string}
        purchaseDate: {type: string, format: date-time}
    AssetCollection:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/Asset'}
...